test-pipeline

+++ scripts/stop_server.sh
@@ -3,8 +3,18 @@
 #source ~/.bashrc
 #catalina.sh stop
 #
-sudo killall java
-##exit 0
+#!/bin/bash
+
+# Kiểm tra xem có quy trình Java đang chạy không
+if pgrep -x "java" > /dev/null
+then
+  echo "Đang có quy trình Java chạy. Đang tiêu diệt..."
+  sudo killall java
+  echo "Đã tiêu diệt quy trình Java."
+else
+  echo "Không có quy trình Java đang chạy."
+fi
+
 #
 ##rm -r /opt/tomcat/webapps/ROOT
 ##rm /opt/tomcat/webapps/ROOT.war
